EquityView - A React app for monitoring stock market prices
EquityView is a web application that allows users to monitor stock market prices in real time. Built using React on the front-end and a practice server provided by SoftUni on the back-end, EquityView provides users with up-to-date information on the stock market. This application was built with the goal of providing an intuitive and user-friendly experience for investors and traders alike.