Replacing local repository with contents from main branch
Due to an error in my local repository on my computer, I deleted all the files and folders from it. I then downloaded the contents of the "main" branch from GitHub and replaced the deleted files and folders with them. This allowed me to sync my local repository with the "main" branch on GitHub.